# src/main/java/com/example/eventledger/spec/openapi.yaml
openapi: 3.0.3
info:
  title: Event Ledger API
  description: API for managing financial transaction events with idempotency support
  version: 1.0.0
  contact:
    name: API Support

servers:
  - url: http://localhost:8080
    description: Local development server

tags:
  - name: Events
    description: Event management operations
  - name: Accounts
    description: Account balance operations

paths:
  /events:
    post:
      tags:
        - Events
      operationId: submitEvent
      summary: Submit a transaction event
      description: |
        Submit a new event. If eventId already exists, returns existing event (idempotent).
        - Returns 201 Created for new events
        - Returns 200 OK for duplicate events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventRequest'
      responses:
        '201':
          description: Event created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '200':
          description: Duplicate event - returning existing event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Events
      operationId: getEventsByAccount
      summary: List events for an account
      description: List events for an account, ordered by eventTimestamp ASC, then eventId ASC
      parameters:
        - name: account
          in: query
          required: true
          description: Account ID
          schema:
            type: string
        - name: page
          in: query
          required: false
          description: Page number (0-based)
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          description: Page size
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Events retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventResponse'

  /events/{id}:
    get:
      tags:
        - Events
      operationId: getEventById
      summary: Get event by ID
      description: Retrieve a single event by its eventId
      parameters:
        - name: id
          in: path
          required: true
          description: Event ID
          schema:
            type: string
      responses:
        '200':
          description: Event found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /accounts/{accountId}/balance:
    get:
      tags:
        - Accounts
      operationId: getAccountBalance
      summary: Get account balance
      description: |
        Return computed balance for account.
        Balance = sum(CREDIT amounts) - sum(DEBIT amounts)
      parameters:
        - name: accountId
          in: path
          required: true
          description: Account ID
          schema:
            type: string
      responses:
        '200':
          description: Balance retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'

components:
  schemas:
    EventRequest:
      type: object
      required:
        - eventId
        - accountId
        - type
        - amount
        - currency
        - eventTimestamp
      properties:
        eventId:
          type: string
          description: Unique event identifier
          example: evt-001
        accountId:
          type: string
          description: Account identifier
          example: acct-123
        type:
          $ref: '#/components/schemas/EventType'
        amount:
          type: number
          format: decimal
          minimum: 0
          exclusiveMinimum: true
          description: Transaction amount (must be greater than 0)
          example: 150.00
        currency:
          type: string
          description: Currency code
          example: USD
        eventTimestamp:
          type: string
          format: date-time
          description: Event timestamp in ISO-8601 format
          example: '2026-05-15T14:02:11Z'
        metadata:
          type: object
          additionalProperties: true
          description: Optional metadata
          example:
            source: mainframe-batch
            batchId: B-9042

    EventResponse:
      type: object
      properties:
        eventId:
          type: string
          description: Unique event identifier
        accountId:
          type: string
          description: Account identifier
        type:
          $ref: '#/components/schemas/EventType'
        amount:
          type: number
          format: decimal
          description: Transaction amount
        currency:
          type: string
          description: Currency code
        eventTimestamp:
          type: string
          format: date-time
          description: Event timestamp
        metadata:
          type: object
          additionalProperties: true
          description: Optional metadata
        receivedAt:
          type: string
          format: date-time
          description: Timestamp when event was received

    BalanceResponse:
      type: object
      properties:
        accountId:
          type: string
          description: Account identifier
        balance:
          type: number
          format: decimal
          description: Current account balance
        currency:
          type: string
          nullable: true
          description: Currency code (null if no events exist)

    EventType:
      type: string
      enum:
        - CREDIT
        - DEBIT
      description: Type of transaction

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
        status:
          type: integer
          description: HTTP status code
        errorCode:
          type: string
          description: Custom error code
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Error message
        path:
          type: string
          description: Request path
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'

    FieldError:
      type: object
      properties:
        field:
          type: string
          description: Field name
        message:
          type: string
          description: Error message for the field